RKKPLN-1234567

+++ gdp_forecast/cash.cpp
@@ -6,15 +6,15 @@
 
 #pragma once
 
-void cashquoter(int32_t quot, vector<signed char>& TheoWeights, vector<vector<signed char>>& indicatDateInc, vector<vector<signed char>>& Cash1)
+void cashquoter(int64_t quot, vector<signed char>& TheoWeights, vector<vector<signed char>>& indicatDateInc, vector<vector<signed char>>& Cash1)
 {
 	Signch obj1;
 	
-	int32_t t2 = indicatDateInc[0].size() - 1;
+	int64_t t2 = indicatDateInc[0].size() - 1;
 	if (quot > t2) { return; }
-	int32_t t1 = indicatDateInc.size();
+	int64_t t1 = indicatDateInc.size();
 
-	for (int32_t i = 0; i < t1; i++)
+	for (int64_t i = 0; i < t1; i++)
 	{
 		signed char t3 = TheoWeights[i];
         signed char t4 = indicatDateInc[i][quot];
@@ -28,8 +28,8 @@ void cashquoter(int32_t quot, vector<signed char>& TheoWeights, vector<vector<si
 
 void cashall(vector<signed char>& TheoWeights, vector<vector<signed char>>& indicatDateInc, vector<vector<signed char>>& Cash1)
 {
-	int32_t quot = 0;
-	int32_t t0 = indicatDateInc[0].size();
+	int64_t quot = 0;
+	int64_t t0 = indicatDateInc[0].size();
 	while (quot < t0)
 
 	{
